[refactor] package 이름 수정
