+++ src/main/java/kr/startoff/backend/controller/AuthController.java
@@ -3,7 +3,6 @@ package kr.startoff.backend.controller;
 import java.net.URI;
 import java.util.Optional;
 
-import javax.security.auth.login.FailedLoginException;
 import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -16,7 +15,6 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -27,9 +25,9 @@ import kr.startoff.backend.entity.User;
 import kr.startoff.backend.exception.EmailOrNicknameDuplicateException;
 import kr.startoff.backend.exception.InvalidPasswordException;
 import kr.startoff.backend.exception.TokenRefreshException;
-import kr.startoff.backend.model.request.LoginRequest;
-import kr.startoff.backend.model.request.SignupRequest;
-import kr.startoff.backend.model.response.JwtResponse;
+import kr.startoff.backend.payload.request.LoginRequest;
+import kr.startoff.backend.payload.request.SignupRequest;
+import kr.startoff.backend.payload.response.JwtResponse;
 import kr.startoff.backend.security.UserPrincipal;
 import kr.startoff.backend.security.jwt.JwtUtil;
 import kr.startoff.backend.service.UserDetailsServiceImpl;
