+++ src/main/java/kr/startoff/backend/payload/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package kr.startoff.backend.model.request;
+package kr.startoff.backend.payload.request;
 
 import javax.validation.constraints.Email;
 import javax.validation.constraints.NotBlank;
@@ -6,12 +6,10 @@ import javax.validation.constraints.NotBlank;
 import lombok.Getter;
 
 @Getter
-public class SignupRequest {
+public class LoginRequest {
 	@NotBlank
 	@Email
 	private String email;
-	@NotBlank
-	private String nickname;
 	@NotBlank
 	private String password;
 }
