+++ src/main/java/kr/startoff/backend/service/UserService.java
@@ -7,10 +7,10 @@ import org.springframework.transaction.annotation.Transactional;
 import kr.startoff.backend.entity.AuthProvider;
 import kr.startoff.backend.entity.User;
 import kr.startoff.backend.exception.UserNotFoundException;
-import kr.startoff.backend.model.request.SignupRequest;
-import kr.startoff.backend.model.request.UserInfoUpdateRequest;
-import kr.startoff.backend.model.response.UserInfoResponse;
-import kr.startoff.backend.model.response.UserProfileResponse;
+import kr.startoff.backend.payload.request.SignupRequest;
+import kr.startoff.backend.payload.request.UserInfoUpdateRequest;
+import kr.startoff.backend.payload.response.UserInfoResponse;
+import kr.startoff.backend.payload.response.UserProfileResponse;
 import kr.startoff.backend.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
